Сделали репозиторий с EF и подключили его.

+++ MusicShop.Bussines/Features/Albums/Delete/AlbumDeleteRequest.cs
@@ -4,7 +4,7 @@ namespace MusicShop.Bussines.Features.Albums.Delete
 {
     public class AlbumDeleteRequest
     {
-        [Range(1, long.MaxValue, ErrorMessage = "Введено неверное значение")]
-        public long AlbumId { get; set; }
+        [Range(1, int.MaxValue, ErrorMessage = "Введено неверное значение")]
+        public int AlbumId { get; set; }
     }
 }
