+++ MusicShop.Bussines/Features/Artists/Get/ArtistGetRequest.cs
@@ -4,7 +4,7 @@ namespace MusicShop.Bussines.Features.Artists.Get
 {
     public class ArtistGetRequest
     {
-        [Range(1, long.MaxValue, ErrorMessage = "Введено неверное значение")]
-        public long ArtistId { get; set; }
+        [Range(1, int.MaxValue, ErrorMessage = "Введено неверное значение")]
+        public int ArtistId { get; set; }
     }
 }
