+++ MusicShop.Domain/Repositories/IAlbumRepository.cs
@@ -4,12 +4,12 @@ namespace MusicShop.Data
 {
     public interface IAlbumRepository
     {
-        Task<bool> DeleteAlbum(long albumId);
-        Task<Album?> Get(long albumId);
+        Task<bool> DeleteAlbum(int albumId);
+        Task<Album?> Get(int albumId);
         Task<IEnumerable<Album>> GetAll();
-        Task<long?> InsertAlbum(string title, long artistId);
+        Task<int?> InsertAlbum(string title, int artistId);
         Task<IEnumerable<Album>> Search(string titleSearch);
-        Task<bool> UpdateAlbum(long albumId, string title, long artistId);
-        Task<bool> AlbumIsExist(string title, long artistId);
+        Task<bool> UpdateAlbum(int albumId, string title, int artistId);
+        Task<bool> AlbumIsExist(string title, int artistId);
     }
 }
