+++ MusicShop.Domain/Repositories/IArtistRepository.cs
@@ -4,12 +4,12 @@ namespace MusicShop.Data
 {
     public interface IArtistRepository
     {
-        Task<bool> DeleteArtist(long artistId);
-        Task<Artist?> Get(long artistId);
+        Task<bool> DeleteArtist(int artistId);
+        Task<Artist?> Get(int artistId);
         Task<IEnumerable<Artist>> GetAll();
-        Task<long?> InsertArtist(string name);
+        Task<int?> InsertArtist(string name);
         Task<IEnumerable<Artist>> Search(string nameSearch);
-        Task<bool> UpdateArtist(long artistId, string name);
+        Task<bool> UpdateArtist(int artistId, string name);
         Task<bool> ArtistIsExist(string name);
     }
 }
